+++ rustfst/src/fst_impls/const_fst/mod.rs
@@ -1,11 +1,10 @@
 pub use self::data_structure::ConstFst;
 pub(crate) use self::data_structure::ConstState;
 
-mod arc_iterator;
+mod iterators;
 mod converters;
 mod data_structure;
 mod expanded_fst;
 mod fst;
 mod misc;
-mod state_iterator;
 mod text_parser;