+++ rustfst/src/fst_impls/vector_fst/mod.rs
@@ -1,12 +1,11 @@
 pub use self::data_structure::VectorFst;
 pub(crate) use self::data_structure::VectorFstState;
 
-mod arc_iterator;
 mod data_structure;
 mod expanded_fst;
 mod fst;
 mod misc;
 mod mutable_fst;
-mod state_iterator;
+mod iterators;
 mod test;
 mod text_parser;